* changes:
  fix(st): cast clk_get_rate return when setting console
  fix(st): cast the read_cntfrq_el0() return to unsigned int
  fix(st-clock): correct some STM32MP15 clock driver function
  fix(st-pmic): store regulator ID as an uint8_t
  fix(st-regulator): remove const for driver_data
  fix(st-pmic): stub some functions for STPMIC1L
  fix(st-sdmmc2): update plat_sdmmc2_use_dma prototype
  fix(st-drivers): use BIT_32/GENMASK_32 in ST drivers

+++ drivers/st/crypto/stm32_rng.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2025, STMicroelectronics - All Rights Reserved
+ * Copyright (c) 2022-2026, STMicroelectronics - All Rights Reserved
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -320,14 +320,14 @@ int stm32_rng_init(void)
 
 		if (dt_rng.reset >= 0) {
 
-			ret = stm32mp_reset_assert((unsigned long)dt_rng.reset, TIMEOUT_US_1MS);
+			ret = stm32mp_reset_assert((uint32_t)dt_rng.reset, TIMEOUT_US_1MS);
 			if (ret != 0) {
 				panic();
 			}
 
 			udelay(20);
 
-			ret = stm32mp_reset_deassert((unsigned long)dt_rng.reset, TIMEOUT_US_1MS);
+			ret = stm32mp_reset_deassert((uint32_t)dt_rng.reset, TIMEOUT_US_1MS);
 			if (ret != 0) {
 				panic();
 			}
